fix & remove:변경된 댓글수정 로직 반영 및 기존 프로필 수정로직 제거
